requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
langchain==0.3.19
langchain-community==0.3.18
langchain-groq==0.3.2
langchain-huggingface==0.1.2
chromadb==0.6.3
pypdf==5.4.0
python-dotenv==1.0.1
streamlit==1.44.0
aiohttp==3.11.14
fpdf2==2.8.3
sentence-transformers==3.4.1
livekit>=0.20.0
livekit-api>=1.0.2
livekit-agents>=1.0.21
livekit-plugins-openai>=1.5.0
livekit-plugins-cartesia>=1.5.1